--- a/.github/workflows/label.yml
+++ b/.github/workflows/label.yml
@@ -10,10 +10,9 @@ on: [pull_request]
 
 jobs:
   label:
-
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/labeler@v2
-      with:
-        repo-token: "${{ secrets.GITHUB_TOKEN }}"
+      - uses: actions/labeler@v2
+        with:
+          repo-token: "${{ secrets.GITHUB_TOKEN }}"