--- a/packages/avatar/package.json
+++ b/packages/avatar/package.json
@@ -9,7 +9,9 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
-  "sideEffects": ["dist/**/*"],
+  "sideEffects": [
+    "dist/**/*"
+  ],
   "files": [
     "es/*",
     "lib/*",