--- a/packages/badge/package.json
+++ b/packages/badge/package.json
@@ -9,6 +9,7 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
+  "sideEffects": ["dist/**/*"],
   "files": [
     "es/*",
     "lib/*",