--- a/packages/elevation/package.json
+++ b/packages/elevation/package.json
@@ -9,6 +9,7 @@
   "main": "./dist/scssVariables.js",
   "module": "./dist/scssVariables.js",
   "types": "./dist/scssVariables.d.ts",
+  "sideEffects": ["dist/**/*"],
   "files": [
     "src/*",
     "dist/*"