--- a/packages/elevation/package.json
+++ b/packages/elevation/package.json
@@ -9,7 +9,9 @@
   "main": "./dist/scssVariables.js",
   "module": "./dist/scssVariables.js",
   "types": "./dist/scssVariables.d.ts",
-  "sideEffects": ["dist/**/*"],
+  "sideEffects": [
+    "dist/**/*"
+  ],
   "files": [
     "src/*",
     "dist/*"