--- a/packages/form/package.json
+++ b/packages/form/package.json
@@ -10,7 +10,9 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
-  "sideEffects": ["dist/**/*"],
+  "sideEffects": [
+    "dist/**/*"
+  ],
   "files": [
     "es/*",
     "lib/*",