--- a/packages/link/package.json
+++ b/packages/link/package.json
@@ -9,7 +9,9 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
-  "sideEffects": ["dist/**/*"],
+  "sideEffects": [
+    "dist/**/*"
+  ],
   "files": [
     "es/*",
     "lib/*",
@@ -46,7 +48,7 @@
     "classnames": "^2.2.6"
   },
   "devDependencies": {
-    "@reach/router": "^1.3.3",
+    "@reach/router": "^1.3.4",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0"