--- a/packages/material-icons/package.json
+++ b/packages/material-icons/package.json
@@ -14,6 +14,7 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
+  "sideEffects": ["dist/**/*"],
   "files": [
     "es/*",
     "lib/*",