--- a/packages/material-icons/package.json
+++ b/packages/material-icons/package.json
@@ -14,7 +14,9 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
-  "sideEffects": ["dist/**/*"],
+  "sideEffects": [
+    "dist/**/*"
+  ],
   "files": [
     "es/*",
     "lib/*",