--- a/packages/media/package.json
+++ b/packages/media/package.json
@@ -9,6 +9,7 @@
   "main": "./lib/index.js",
   "module": "./es/index.js",
   "types": "./types/index.d.ts",
+  "sideEffects": ["dist/**/*"],
   "files": [
     "es/*",
     "lib/*",