--- a/packages/progress/LICENSE
+++ b/packages/progress/LICENSE
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) 2015-2020 Mikkel Laursen
+Copyright (c) 2015 Mikkel Laursen
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal