--- a/packages/react-md/package.json
+++ b/packages/react-md/package.json
@@ -11,7 +11,9 @@
   "module": "./es/index.js",
   "unpkg": "./dist/umd/react-md.production.min.js",
   "types": "./types/index.d.ts",
-  "sideEffects": ["dist/**/*"],
+  "sideEffects": [
+    "dist/**/*"
+  ],
   "files": [
     "es/*",
     "lib/*",
@@ -74,11 +76,11 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^13.0.0",
-    "@rollup/plugin-node-resolve": "^8.0.1",
+    "@rollup/plugin-node-resolve": "^8.1.0",
     "@rollup/plugin-replace": "^2.3.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "rollup": "^2.15.0",
+    "rollup": "^2.18.0",
     "rollup-plugin-terser": "^6.1.0"
   },
   "peerDependencies": {